removed errorDefinitions file, and placed into errors folder instead. Added class variable in stores/protocol to allow specifying error responses related directly to loading/unloading the Hkl object store
