+++ src/diffcalc_API/server.py
@@ -1,15 +1,15 @@
 from diffcalc.util import DiffcalcException
 from fastapi import Depends, FastAPI, Request, responses
 
-from diffcalc_API import errorDefinitions
 from diffcalc_API.errors.Constraints import responses as responsesConstraints
+from diffcalc_API.errors.definitions import DiffcalcAPIException
 from diffcalc_API.errors.HklCalculation import responses as responsesHkl
 from diffcalc_API.errors.UBCalculation import responses as responsesUb
 from diffcalc_API.stores.pickling import get_store
 
 from . import routes
 
-app = FastAPI(responses=errorDefinitions.responses)
+app = FastAPI(responses=get_store().responses)
 
 app.include_router(routes.UBCalculation.router, responses=responsesUb)
 app.include_router(routes.Constraints.router, responses=responsesConstraints)
@@ -28,10 +28,8 @@ async def diffcalc_exception_handler(request: Request, exc: DiffcalcException):
     )
 
 
-@app.exception_handler(errorDefinitions.DiffcalcAPIException)
-async def http_exception_handler(
-    request: Request, exc: errorDefinitions.DiffcalcAPIException
-):
+@app.exception_handler(DiffcalcAPIException)
+async def http_exception_handler(request: Request, exc: DiffcalcAPIException):
     return responses.JSONResponse(
         status_code=exc.status_code,
         content={"message": exc.detail, "type": str(type(exc))},
