+++ src/diffcalc_API/stores/pickling.py
@@ -1,13 +1,47 @@
 import pickle
 from pathlib import Path
 
+import numpy as np
 from diffcalc.hkl.calc import HklCalculation
 from diffcalc.hkl.constraints import Constraints
 from diffcalc.ub.calc import UBCalculation
 
 from diffcalc_API.config import savePicklesFolder
-from diffcalc_API.errorDefinitions import attempting_to_overwrite, check_file_exists
+from diffcalc_API.errors.definitions import (
+    DiffcalcAPIException,
+    ErrorCodes,
+    allResponses,
+)
 from diffcalc_API.stores.protocol import HklCalcStore
+
+
+class Codes(ErrorCodes):
+    attempting_to_overwrite = 405
+    check_file_exists = 404
+
+
+def attempting_to_overwrite(fileName: str) -> None:
+    pickledFile = Path(savePicklesFolder) / fileName
+    if (pickledFile).is_file():
+        errorMessage = (
+            f"File already exists for crystal {fileName}!"
+            f"\nEither delete via DELETE request to this URL "
+            f"or change the existing properties. "
+        )
+        raise DiffcalcAPIException(status_code=405, detail=errorMessage)
+
+    return
+
+
+def check_file_exists(pickledFile: Path, name: str) -> None:
+    if not (pickledFile).is_file():
+        errorMessage = (
+            f"File for crystal {name} not found."
+            f"\nYou need to post to"
+            f" http://localhost:8000/{name}"
+            f" first to generate the pickled file.\n"
+        )
+        raise DiffcalcAPIException(status_code=404, detail=errorMessage)
 
 
 class PicklingHklCalcStore:
@@ -15,6 +49,9 @@ class PicklingHklCalcStore:
 
     def __init__(self, root_directory: Path) -> None:
         self._root_directory = root_directory
+        self.responses = {
+            code: allResponses[code] for code in np.unique(Codes().all_codes())
+        }
 
     async def create(self, name: str) -> None:
         attempting_to_overwrite(name)
