+++ src/diffcalc_API/stores/protocol.py
@@ -1,4 +1,4 @@
-from typing import Protocol
+from typing import Any, Dict, Protocol, Union
 
 from diffcalc.hkl.calc import HklCalculation
 
@@ -7,6 +7,8 @@ class HklCalcStore(Protocol):
     """
     Protocol, or interface, for interacting with the Hkl object.
     """
+
+    responses: Dict[Union[int, str], Dict[str, Any]]
 
     async def create(self, name: str) -> None:
         ...
